refactor: create ambient_message
